+++ KenkaiSDKCore/Sources/Utility/Constants/CoreConstants.swift
@@ -46,7 +46,7 @@ public class CoreConstants {
         return contentBlock.rawValue
     }
 
-    var SDKVersion: String = "ios/1.0.0"
+    var SDKVersion: String = "ios/1.0.1"
 
     public var updateImmediately: Bool = false