+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Documenting the changes in the SDK subject to each release version.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## **[1.0.1]  - 10.11.2025**
+
+- Include `CFCoreEvent.shared.fetchActions(...)"` for fetching any type of actions with the relevant params
+- Update Action callbacks to return all attributes from the action request
+- Default include `notification_permission` in user catalog/dimension with default values as `granted` & `denied`
+
+
 ## **[1.0.0]  - 31.10.2025**
 
 - Initial SDK Release